방명록 작성

+++ server/routes/visitorRoutes.js
@@ -1,16 +1,21 @@
 const express = require("express");
+const bcrypt = require("bcrypt");
 const router = express.Router();
 const Visitor = require("../models/Visitor");
 
 router.post("/add-visitor", async (req, res) => {
   try {
-    const { visitor_no, writer, writer_avatar, content } = req.body;
+    const { visitor_no, writer, writer_avatar, content, password } = req.body;
+
+    const salt = await bcrypt.genSalt(10);
+    const hashedPassword = await bcrypt.hash(password, salt);
 
     const newVisitor = new Visitor({
       visitor_no,
       writer,
       writer_avatar,
       content,
+      password: hashedPassword, 
     });
 
     await newVisitor.save();
@@ -48,13 +53,21 @@ router.get("/visitor/:id", async (req, res) => {
 router.put("/visitor-update/:id", async (req, res) => {
   try {
     const { id } = req.params;
-    const { visitor_no, writer, writer_avatar, content } = req.body;
+    const { visitor_no, writer, writer_avatar, content, currentPassword } = req.body;
 
-    const updatedVisitor = await Visitor.findByIdAndUpdate(id, { visitor_no, writer, writer_avatar, content }, { new: true });
-
-    if (!updatedVisitor) {
+    const existingVisitor = await Visitor.findById(id);
+    if (!existingVisitor) {
       return res.status(404).json({ message: "Visitor not found" });
     }
+
+    const isPasswordCorrect = await bcrypt.compare(currentPassword, existingVisitor.password);
+    if (!isPasswordCorrect) {
+      return res.status(403).json({ message: "Incorrect password" }); 
+    }
+
+    const updatedFields = { visitor_no, writer, writer_avatar, content };
+
+    const updatedVisitor = await Visitor.findByIdAndUpdate(id, updatedFields, { new: true });
 
     res.status(200).json({ message: "Visitor successfully updated", visitor: updatedVisitor });
   } catch (error) {
@@ -65,12 +78,19 @@ router.put("/visitor-update/:id", async (req, res) => {
 router.delete("/visitor-delete/:id", async (req, res) => {
   try {
     const { id } = req.params;
+    const { currentPassword } = req.body; 
 
-    const deletedVisitor = await Visitor.findByIdAndDelete(id);
-
-    if (!deletedVisitor) {
+    const existingVisitor = await Visitor.findById(id);
+    if (!existingVisitor) {
       return res.status(404).json({ message: "Visitor not found" });
     }
+
+    const isPasswordCorrect = await bcrypt.compare(currentPassword, existingVisitor.password);
+    if (!isPasswordCorrect) {
+      return res.status(403).json({ message: "Incorrect password" }); 
+    }
+
+    await Visitor.findByIdAndDelete(id);
 
     res.status(200).json({ message: "Visitor successfully deleted" });
   } catch (error) {
